clean up pgxcommon
we used a "transaction factory" everywhere but never used it to create a transaction (ever since we switched to implicit transactions).

this removes the extra abstraction. it also gives the DBReader interface
a better name (since it can be used for more than just reading)

+++ internal/datastore/crdb/options.go
@@ -15,7 +15,7 @@ type crdbOptions struct {
 	followerReadDelay           time.Duration
 	maxRevisionStalenessPercent float64
 	gcWindow                    time.Duration
-	maxRetries                  uint32
+	maxRetries                  uint8
 	splitAtUsersetCount         uint16
 	overlapStrategy             string
 	overlapKey                  string
@@ -259,7 +259,7 @@ func GCWindow(window time.Duration) Option {
 // MaxRetries is the maximum number of times a retriable transaction will be
 // client-side retried.
 // Default: 5
-func MaxRetries(maxRetries uint32) Option {
+func MaxRetries(maxRetries uint8) Option {
 	return func(po *crdbOptions) { po.maxRetries = maxRetries }
 }
